Revert "Revert "Add[PlayerAttack]""
This reverts commit d9b7e6723c054ce01751054a45560551410d18ff.
