+++ Assets/Scripts/LastPlayer/LastPlayerController.cs
@@ -8,6 +8,7 @@ public class LastPlayerController : MonoBehaviour
     private Animator anim;
     private Rigidbody2D rb;
     private GameManager gameManager;
+    private PlayerAttack playerAttack;
 
     public FadeOut fadeOut;
     public CharacterStats characterStats;
@@ -52,20 +53,20 @@ public class LastPlayerController : MonoBehaviour
     [SerializeField] private float currentStamina;
     [SerializeField] private float staminaRegenRate = 10f;
     [SerializeField] private float dashStaminaCost = 20f;
-    [SerializeField] private float attackStaminaCost = 5f;
-    [SerializeField] private float comboStaminaCost = 20f;
+    [SerializeField] public float attackStaminaCost = 5f;
+    [SerializeField] public float comboStaminaCost = 20f;
 
-    public Transform attackPoint;
-    [SerializeField] private float attackRange = 1f;
-    [SerializeField] private LayerMask enemyLayer;
+    //public Transform attackPoint;
+    //[SerializeField] private float attackRange = 1f;
+    //[SerializeField] private LayerMask enemyLayer;
 
-    private float lastAttackTime = 0f;
-    public float attackRate = 1f;
-    float nextAttackTime = 0f;
-    private int attackClickCount = 1;
+    //private float lastAttackTime = 0f;
+    //public float attackRate = 1f;
+    //float nextAttackTime = 0f;
+    //private int attackClickCount = 1;
 
-    public bool canTakeDamage = true;
-    private int damage = 10;
+    //public bool canTakeDamage = true;
+    //private int damage = 10;
 
 
     [HideInInspector] public bool ledgeDetected;
@@ -109,10 +110,10 @@ public class LastPlayerController : MonoBehaviour
             isWallSliding = true;
             rb.velocity = new Vector2(rb.velocity.x, rb.velocity.y * 0.1f);
         }
-        if (Time.time > lastAttackTime + 5f)
-        {
-            attackClickCount = 1;
-        }
+        //if (Time.time > lastAttackTime + 5f)
+        //{
+        //    attackClickCount = 1;
+        //}
         if (isLadderDetected)
         {
             ClimbLadder();
@@ -123,7 +124,7 @@ public class LastPlayerController : MonoBehaviour
             Move();
 
             Dash();
-            CheckAttackTime();
+            //CheckAttackTime();
             CheckForLedge();
         }
         Death();
@@ -178,17 +179,17 @@ public class LastPlayerController : MonoBehaviour
     }
 
 
-    private void CheckAttackTime()
-    {
-        if (Time.time >= nextAttackTime)
-        {
-            if (Input.GetMouseButtonDown(0) && isGrounded && PopupUIManager.instance.activePopupLList.Count <= 0)
-            {
-                nextAttackTime = Time.time + 0.5f / attackRate;
-                Attack();
-            }
-        }
-    }
+    //private void CheckAttackTime()
+    //{
+    //    if (Time.time >= nextAttackTime)
+    //    {
+    //        if (Input.GetMouseButtonDown(0) && isGrounded && PopupUIManager.instance.activePopupLList.Count <= 0)
+    //        {
+    //            nextAttackTime = Time.time + 0.5f / attackRate;
+    //            Attack();
+    //        }
+    //    }
+    //}
     public void CheckInput()
     {
         movingInput = Input.GetAxis("Horizontal");
@@ -199,12 +200,12 @@ public class LastPlayerController : MonoBehaviour
         }
         if (Input.GetMouseButtonDown(1))
         {
-            canTakeDamage = false;
+            playerAttack.canTakeDamage = false;
             Debug.Log("누름");
         }
         else if (Input.GetMouseButtonUp(1))
         {
-            canTakeDamage = true;
+            playerAttack.canTakeDamage = true;
             Debug.Log("땜");
         }
         if (Input.GetKeyDown(KeyCode.Space))
@@ -277,73 +278,73 @@ public class LastPlayerController : MonoBehaviour
     //            ApplyDamage(modifiedAttackDamage);
     //        }
     //}
-    private void Attack()
-    {
-        if (gameManager.playerStats.characterStamina >= attackStaminaCost)
-        {
-            gameManager.playerStats.characterStamina -= attackStaminaCost;
-            anim.SetTrigger("attack");
-            gameManager.playerStats.AttackDamage(damage);
-            int modifiedAttackDamage = damage;
-            if (attackClickCount != 0 && attackClickCount % 3 == 0)
-            {
-                gameManager.playerStats.characterStamina -= comboStaminaCost;
-                anim.SetTrigger("combo");
-                modifiedAttackDamage += 10;
-                attackClickCount = 0;
-            }
-            ApplyDamage(modifiedAttackDamage);
-        }
-    }
+    //private void Attack()
+    //{
+    //    if (gameManager.playerStats.characterStamina >= attackStaminaCost)
+    //    {
+    //        gameManager.playerStats.characterStamina -= attackStaminaCost;
+    //        anim.SetTrigger("attack");
+    //        gameManager.playerStats.AttackDamage(damage);
+    //        int modifiedAttackDamage = damage;
+    //        if (attackClickCount != 0 && attackClickCount % 3 == 0)
+    //        {
+    //            gameManager.playerStats.characterStamina -= comboStaminaCost;
+    //            anim.SetTrigger("combo");
+    //            modifiedAttackDamage += 10;
+    //            attackClickCount = 0;
+    //        }
+    //        ApplyDamage(modifiedAttackDamage);
+    //    }
+    //}
 
-    //Debug.Log(canTakeDamage);//PlayerToMonster
+    ////Debug.Log(canTakeDamage);//PlayerToMonster
 
-    private void ApplyDamage(int damage) // Add damage To Monster
-    {
-        Collider2D[] hitEnemies = Physics2D.OverlapCircleAll(attackPoint.position, attackRange, enemyLayer);
-        foreach (Collider2D enemyCollider in hitEnemies)
-        {
-            if (enemyCollider.CompareTag("Boss_DB"))
-            {
-                lastAttackTime = Time.time;
-                attackClickCount++;
-                DeathBringerEnemy deathBringer = enemyCollider.GetComponent<DeathBringerEnemy>();
-                if (deathBringer != null)
-                {
-                    deathBringer.TakeDamage(damage);
-                    PlayerEvents.playerDamaged.Invoke(gameObject, damage);
-                }
-            }
-            else if (enemyCollider.CompareTag("Boss_Archer"))
-            {
-                Boss_Archer boss_archer = enemyCollider.GetComponent<Boss_Archer>();
-                if (boss_archer != null)
-                {
-                    Debug.Log("Deal" + characterStats.characterNomallAttackDamage + " damage to Boss Archer.");
-                    //boss_archer.TakeDamage(attackDamage);
-                    boss_archer.TakeDamage(characterStats.characterNomallAttackDamage);
-                }
-            }
-            else if (enemyCollider.CompareTag("skeleton"))
-            {
-                skeletonEnemy skeleton = enemyCollider.GetComponent<skeletonEnemy>();
-                if (skeleton != null)
-                {
-                    Debug.Log("Deal" + characterStats.characterNomallAttackDamage + " damage to Skeleton.");
-                    skeleton.TakeDamage(characterStats.characterNomallAttackDamage);
-                }
-            }
-            else if (enemyCollider.CompareTag("archer"))
-            {
-                archerEnemy archer = enemyCollider.GetComponent<archerEnemy>();
-                if (archer != null)
-                {
-                    Debug.Log("Deal " + characterStats.characterNomallAttackDamage + " damage to Archer.");
-                    archer.TakeDamage(characterStats.characterNomallAttackDamage);
-                }
-            }
-        }
-    }
+    //private void ApplyDamage(int damage) // Add damage To Monster
+    //{
+    //    Collider2D[] hitEnemies = Physics2D.OverlapCircleAll(attackPoint.position, attackRange, enemyLayer);
+    //    foreach (Collider2D enemyCollider in hitEnemies)
+    //    {
+    //        if (enemyCollider.CompareTag("Boss_DB"))
+    //        {
+    //            lastAttackTime = Time.time;
+    //            attackClickCount++;
+    //            DeathBringerEnemy deathBringer = enemyCollider.GetComponent<DeathBringerEnemy>();
+    //            if (deathBringer != null)
+    //            {
+    //                deathBringer.TakeDamage(damage);
+    //                PlayerEvents.playerDamaged.Invoke(gameObject, damage);
+    //            }
+    //        }
+    //        else if (enemyCollider.CompareTag("Boss_Archer"))
+    //        {
+    //            Boss_Archer boss_archer = enemyCollider.GetComponent<Boss_Archer>();
+    //            if (boss_archer != null)
+    //            {
+    //                Debug.Log("Deal" + characterStats.characterNomallAttackDamage + " damage to Boss Archer.");
+    //                //boss_archer.TakeDamage(attackDamage);
+    //                boss_archer.TakeDamage(characterStats.characterNomallAttackDamage);
+    //            }
+    //        }
+    //        else if (enemyCollider.CompareTag("skeleton"))
+    //        {
+    //            skeletonEnemy skeleton = enemyCollider.GetComponent<skeletonEnemy>();
+    //            if (skeleton != null)
+    //            {
+    //                Debug.Log("Deal" + characterStats.characterNomallAttackDamage + " damage to Skeleton.");
+    //                skeleton.TakeDamage(characterStats.characterNomallAttackDamage);
+    //            }
+    //        }
+    //        else if (enemyCollider.CompareTag("archer"))
+    //        {
+    //            archerEnemy archer = enemyCollider.GetComponent<archerEnemy>();
+    //            if (archer != null)
+    //            {
+    //                Debug.Log("Deal " + characterStats.characterNomallAttackDamage + " damage to Archer.");
+    //                archer.TakeDamage(characterStats.characterNomallAttackDamage);
+    //            }
+    //        }
+    //    }
+    //}
 
 
     private void Death()
@@ -461,11 +462,11 @@ public class LastPlayerController : MonoBehaviour
         Gizmos.DrawLine(transform.position, new Vector3(transform.position.x, transform.position.y - groundCheckDistance));
         Gizmos.DrawLine(transform.position, new Vector3(transform.position.x, transform.position.y + ceilCheckDistance));
 
-        if (attackPoint == null)
-        {
-            return;
-        }
-        Gizmos.DrawWireSphere(attackPoint.position, attackRange);
+        //if (attackPoint == null)
+        //{
+        //    return;
+        //}
+        //Gizmos.DrawWireSphere(attackPoint.position, attackRange);
 
     }
 }
