+++ Assets/Scripts/LastPlayer/PlayerAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 35a77a4c81ff545d48483969bdae174a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
